pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ai-recipe-crew"
version = "0.1.0"
description = "CrewAI-based cooking recipe recommendation microservice"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "crewai>=0.55.0",
    "crewai-tools>=0.8.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "langchain-chroma>=0.1.0",
    "chromadb>=0.5.0",
    "openai>=1.30.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "tenacity>=8.3.0",
]

[tool.setuptools.packages.find]
where = ["src"]
